fix(profile): Refractoring of the user profile screen
