+++ lib/views/navigation/routes.dart
@@ -2,7 +2,7 @@ import "package:flutter/material.dart";
 import "package:proxima/views/pages/create_account_page.dart";
 import "package:proxima/views/pages/home_page.dart";
 import "package:proxima/views/pages/login/login_page.dart";
-import "package:proxima/views/pages/profile_page.dart";
+import "package:proxima/views/pages/profile/profile_page.dart";
 
 enum Routes {
   home("home"),
