+++ lib/views/pages/profile/posts_info/info_card.dart
@@ -1,0 +1,25 @@
+import "package:flutter/material.dart";
+
+class InfoCard extends StatelessWidget {
+  const InfoCard({
+    super.key,
+    required this.theme,
+  });
+
+  final ThemeData theme;
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      width: 92,
+      height: 92,
+      decoration: BoxDecoration(
+        border: Border.all(
+          color: Colors.black,
+          width: 2,
+        ),
+        borderRadius: const BorderRadius.all(Radius.circular(10)),
+      ),
+    );
+  }
+}
