+++ lib/views/pages/profile/posts_info/info_row.dart
@@ -1,0 +1,51 @@
+import "package:flutter/material.dart";
+
+class InfoRow extends StatelessWidget {
+  const InfoRow({
+    super.key,
+    required this.theme,
+    required this.itemList,
+    required this.title,
+  });
+
+  final ThemeData theme;
+  final List<Widget> itemList;
+  final String title;
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      height: 150,
+      width: 380,
+      decoration: BoxDecoration(
+        borderRadius: const BorderRadius.all(Radius.circular(10)),
+        color: theme.colorScheme.secondaryContainer,
+      ),
+      child: Padding(
+        padding: const EdgeInsets.only(left: 5, top: 1),
+        child: Column(
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: [
+            Text(
+              title,
+              style: theme.textTheme.headlineSmall,
+            ),
+            const SizedBox(height: 5),
+            Expanded(
+              child: ListView.separated(
+                padding: const EdgeInsets.all(8),
+                scrollDirection: Axis.horizontal,
+                itemCount: itemList.length,
+                itemBuilder: (BuildContext context, int index) {
+                  return itemList[index];
+                },
+                separatorBuilder: (BuildContext context, int index) =>
+                    const SizedBox(width: 10),
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
